addressing io.py
