+++ ctdcal/scripts/quick_convert.py
@@ -1,7 +1,8 @@
 import logging
 from pathlib import Path
 
-from ctdcal import get_ctdcal_config, io
+from ctdcal import get_ctdcal_config
+from ctdcal.common import load_cnv
 from ctdcal.flagging import flag_common as flagging
 
 log = logging.getLogger(__name__)
@@ -32,7 +33,7 @@ def cnv_to_ct1():
         "CStarTr0": "CTDXMISS",
     }
     for f in cnv_files:
-        df = io.load_cnv(f).rename(mapper=sbe_to_woce, axis=1)
+        df = load_cnv(f).rename(mapper=sbe_to_woce, axis=1)
         df = df[sbe_to_woce.values()]
 
         # give everything WOCE-named uncalibrated flags
